requirements.txt
pysam
pytabix
biopython
gensim
numpy
pandas
scipy
faiss-cpu
tqdm
psutil